requirements.txt
msgpack==1.0.7
msgpack-types==0.2.0
mypy==1.8.0
mypy-extensions==1.0.0
nats-py==2.6.0
protobuf==4.25.2
tomli==2.0.1
typing_extensions==4.9.0
